{{ cookiecutter.app }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.app }}"
version = "0.0.1"
description = "{{ cookiecutter.description }}"
authors = ["{{ cookiecutter.author }} <{{ cookiecutter.author_email }}>"]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
Flask = {extras = ["async"], version = "^2.0.3"}
flask-oidc = "^1.4.0"
flask-babel = "^2.0.0"
flask-httpauth = "^4.4.0"
flask-login = "^0.6.1"
flask-wtf = "^0.15.1"
wtforms = "^2.3.3"
Babel = "^2.9.1"
blinker = "^1.4"
Flask-Bootstrap = "^3.3.7"
Werkzeug = "^2.0.0"
Flask-Moment = "^1.0.2"
python-dotenv = "^0.20.0"
python-editor = "^1.0.4"
python-dateutil = "^2.8.2"
pytz = "^2022.1"
visitor = "^0.1.3"
PyJWT = "^2.4.0"
dominate = "^2.6.0"
alembic = "^1.7.7"
itsdangerous = "^2.1.2"
idna = "^3.3"
greenlet = "^1.1.2"
click = "^8.1.3"
requests = "^2.27.1"
okta-jwt-verifier = "0.2.0"
flask-cors = "^3.0.9"
pyOpenSSL = "^19.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.13.0"
pytest = "^7.1.1"
ipython = "^8.1.1"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
